6. Add modal
