+++ src/components/AdminCardRoom/AdminCardRoom.module.scss
@@ -1,0 +1,6 @@
+.wrap{
+  .img-profile-admin{
+    height: 238px;
+    object-fit: cover;
+  }
+}
