+++ src/components/ModalCustom/ModalCustom.module.scss
@@ -1,0 +1,12 @@
+.wrap{
+  .feature-icon{
+    width: 3rem;
+    height: 3rem;
+    background-color: rgb(89, 89, 255);
+    color: white;
+    border-radius: 0.5rem;
+    justify-content: center;
+    align-items: center;
+    display: flex;
+  }
+}
